Move and rename more files
