--- src/runelite/enum/PlayerDataType.ts
@@ -1,7 +1,0 @@
-enum PlayerDataType {
-  VARBIT = 'varb',
-  VARPLAYER = 'varp',
-  SKILLLEVEL = 'level',
-}
-
-export default PlayerDataType;
