+++ src/runelite/orm/PlayerDataJson.ts
@@ -1,8 +1,8 @@
 import {
   Entity, Column, PrimaryColumn, BaseEntity,
 } from 'typeorm';
-import { ProfileType } from '../runelite/enum/ProfileType';
-import { PlayerDataValue } from '../runelite/types/PlayerDataValue';
+import { ProfileType } from '../enum/ProfileType';
+import { PlayerDataValue } from '../types/PlayerDataValue';
 
 @Entity('player_data_json')
 export default class PlayerDataJson extends BaseEntity {
