+++ src/runelite/transformers/CollectionLogTransformer.ts
