+++ src/runelite/transformers/LeagueTransformer.ts
